+++ odbc-api/src/connection.rs
@@ -16,8 +16,11 @@ impl<'conn> Drop for Connection<'conn> {
                     // Avoid panicking, if we already have a panic. We don't want to mask the original
                     // error.
                     if !panicking() {
-                        panic!("Unexpected error rolling back transaction (In order to recover \
-                                from invalid transaction state during disconnect): {:?}", e)
+                        panic!(
+                            "Unexpected error rolling back transaction (In order to recover \
+                                from invalid transaction state during disconnect): {:?}",
+                            e
+                        )
                     }
                 }
                 // Transaction is rolled back. Now let's try again to disconnect.